@@ -5,13 +5,13 @@ author: null
 homepage: null
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  equatable: ^1.2.5
+  equatable: ^2.0.0-nullsafety.1
 
 dev_dependencies:
   flutter_test:
